Move calculate password strength to util  - Fix Typo  - Move restoreeditor to editor
